# backend/src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Docker Web Desktop API",
    "description": "API documentation for Docker Web Desktop (auto-generated)",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/containers/": {
      "get": {
        "tags": [
          "Containers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "type": "string"
          },
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/containers/{id}": {
      "get": {
        "tags": [
          "Containers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Containers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "force",
            "in": "query",
            "type": "string"
          },
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/containers/{id}/files/download": {
      "get": {
        "tags": [
          "Containers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "path",
            "in": "query",
            "type": "string"
          },
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/compose/file": {
      "post": {
        "tags": [
          "Compose"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "projectName": {
                  "example": "any"
                },
                "workingDir": {
                  "example": "any"
                },
                "configFiles": {
                  "example": "any"
                },
                "source": {
                  "example": "any"
                },
                "wslDistro": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/compose/": {
      "get": {
        "tags": [
          "Compose"
        ],
        "description": "",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}